ref: Move js message handling to a separate file
