+++ ankihub/entry_point.py
@@ -10,7 +10,7 @@ from aqt.gui_hooks import profile_did_open, profile_will_close
 from . import LOGGER
 from .db import ankihub_db
 from .feature_flags import setup_feature_flags_in_background
-from .gui import browser, deckbrowser, editor, progress, reviewer
+from .gui import browser, deckbrowser, editor, js_message_handling, progress, reviewer
 from .gui.addons import setup_addons
 from .gui.auto_sync import setup_auto_sync
 from .gui.config_dialog import setup_config_dialog_manager
@@ -130,6 +130,9 @@ def _general_setup():
 
     setup_addons()
     LOGGER.info("Set up addons.")
+
+    js_message_handling.setup()
+    LOGGER.info("Set up JavaScript message handling.")
 
     setup_config_dialog_manager()
     LOGGER.info("Set up config.")
